+++ Frontend/src/Main.tsx
@@ -1,0 +1,32 @@
+import React from "react";
+import {
+  SafeAreaView,
+  ScrollView,
+  StatusBar,
+  useColorScheme,
+} from "react-native";
+import { Colors } from "react-native/Libraries/NewAppScreen";
+
+import Body from "./components/body/NewBody";
+
+const Main = () => {
+  const isDarkMode = useColorScheme() === "dark";
+
+  const backgroundStyle = {
+    backgroundColor: isDarkMode ? Colors.darker : Colors.lighter,
+  };
+
+  return (
+    <SafeAreaView style={backgroundStyle}>
+      <StatusBar barStyle={isDarkMode ? "light-content" : "dark-content"} />
+      <ScrollView
+        contentInsetAdjustmentBehavior="automatic"
+        style={backgroundStyle}
+      >
+        <Body />
+      </ScrollView>
+    </SafeAreaView>
+  );
+};
+
+export default Main;
